Handle SYSTEM_LOGICAL_PROCESSOR_INFORMATION_EX union as subclassing
The union embedded in SYSTEM_LOGICAL_PROCESSOR_INFORMATION_EX has a
dynamic size. So at decoding time the size depends on the Relationship
code embedded in the structure.

Instead of trying to press this through the default JNA union 
deserialization, model the four cases created by the four embedded
structures in the dummy union as separate subclasses of 
SYSTEM_LOGICAL_PROCESSOR_INFORMATION_EX.